requirements.txt
pandas
psycopg2-binary==2.8.6
python-dotenv==0.19.2
SQLAlchemy==1.3.24
git+https://github.com/JustAnotherArchivist/snscrape.git